Fixed all checkstyle warnings
